requirements.txt
numpy
scipy
matplotlib
Pillow
torch
torchvision
h5py
tensorboard
pycocotools